refactor: extract timeline view models
